underwater_imagery/pyproject.toml
[tool.poetry]
name = "underwater_imagery"
version = "0.1.0"
description = ""
authors = [
    "delimarr <robin.delimar@students.fhnw.ch>",
    "Krabbe  <lukas.krebs@students.fhnw.ch>",
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26"
pillow = "^10.2"
matplotlib = "^3.8"
torch = { url = "https://download.pytorch.org/whl/cu121/torch-2.1.2%2Bcu121-cp311-cp311-win_amd64.whl" }
torchvision = { url = "https://download.pytorch.org/whl/cu121/torchvision-0.16.2%2Bcu121-cp311-cp311-win_amd64.whl" }
torchaudio = { url = "https://download.pytorch.org/whl/cu121/torchaudio-2.1.2%2Bcu121-cp311-cp311-win_amd64.whl" }
torchshow = "^0.5"
ipykernel = "^6.28"


[tool.poetry.dev-dependencies]
isort = "^5.13"
black = "^23.12"
pytest = "^7.4"
pydocstyle = "^6.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
